Drop mysql-common-5.6
This package is now not necessary as each variant can ship its own
my.cnf. Since only binary packages from this source package depended on
it, no transitionary considerations are required; we can just drop the
package.

Mark mysql-common as providing the functionality of mysql-common-5.6 so
that upgrades can be aware and remove mysql-common-5.6 accordingly.

This won't cause a purge of mysql-common-5.6, but this shouldn't matter
as the only configuration mysql-common-5.6 supplied previously was empty
by default. If the user modified it to add real content, the
modifications will remain, which is what we want anyway.