Revert "Cleanup a bit more"
cmake/os/GNU.cmake is added by a quilt patch and this package is format
"3.0 (quilt)". So this will be cleaned up by dpkg-source as necessary.
Cleaning it manually breaks the build on GNU/Hurd since clean is also
run between quilt patch application and the build, nullifying the effect
of the patch.

This reverts commit 7ef4a271973b8dcae1474fa7951681a0e6af4796.

See:

http://lists.alioth.debian.org/pipermail/pkg-mysql-maint/2015-December/008483.html

Thanks to Samuel Thibault for identifying this and for the patch.