d/rules: some dh_install excludes
We want to move to --fail-missing, so skip some known items.

Manpages are installed by dh_installman so we're not sure why they are
being flagged by dh_install.
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=436240#15 suggests
that this isn't just us. Perhaps this is a wishlist item or bug in
dh_install? In the meantime, we can just exclude these.

According to upstream, usr/share/mysql/mysql.server is not needed on
Debian, so is deliberately excluded. This could be moved to
debian/not-installed in the future if the list grows.