Binary installation tweaks
resolve_stack_dump, lz4_decompress and zlib_decompress are shipped in
/usr/lib/mysql after consultation with upstream since MySQL-specific
binaries named like this don't appear suitable for the global namespace.

/usr/share/mysql/docs/INFO_{BIN,SRC} are build information and not
strictly documentation, so belong in /usr/share/mysql. Since upstream
currently ships it in docs/, we do the same for now. The key thing is
that they are correctly in /usr/share/mysql.