Drop d/p/scripts__mysqld_safe.sh__signals.patch
Not so relevant now that we have systemd by default, which no longer
uses mysqld_safe.

Patch is reported as being stale upstream and we cannot test it as we're
using systemd now. If somebody wants to fix this patch and get it
upstreamed and/or maintain it in Debian, please get in touch.