Update translated references from 5.6 to 5.7
I'm deliberately doing this after pulling in all contributed translation
updates. I switch any translated strings that referred to 5.6 to 5.7
instead, which I trust is safe. I also update any references to the
source package name. I deliberately do not change "Project-Id-Version"
metadata as I'm not sure what the consequence is of doing this (for
example to translator tooling).