Removed always-false check on purge in postrm
When purging the packages, users are supposed to be given the option to remove
data files. But they are only given the option if the server binary doesn't
exist. This check seems to be from an old version (4.1) where this logic
was in the top metapackage. Currently this check appears to never pass,
so the code is dead. This can lead to errors if the users want to switch
between MySQL and MariaDB, since the two are no longer data compatible.
Removed the check on the server binary, since it shouldn't be possible for
another package to own that file anyway.
LP: #1602945